- add C-API function Ns_LogSeveritySetEnabled() to let modules change   enabled states - ns_http:     * fix invalid request strings, when url http://google.com/ (with trailing  slash) is used     * fix possible race condition when time gap between "ns_http queue"        and "ns_http wait" is larger - prefer memcpy over strncpy, since it has in most cases the appropriate semantics   and it is faster - avoid multiple strlen() operations - provide better handling of overlong strings - add nonnull assertions

+++ nsd/binder.c
@@ -352,15 +352,24 @@ Ns_SockBindUdp(const struct sockaddr_in *saPtr)
 NS_SOCKET
 Ns_SockBindUnix(const char *path, int socktype, int mode)
 {
-    NS_SOCKET sock;
 #ifdef _WIN32
-    sock = NS_INVALID_SOCKET;
+    return NS_INVALID_SOCKET;
 #else
+    NS_SOCKET sock;
     struct sockaddr_un addr;
+    size_t pathLength;
+
+    assert(path != NULL);
+    pathLength = strlen(path);
+
+    if (pathLength >= sizeof(addr.sun_path)) {
+        Ns_Log(Error, "provided path exeeds maximum length: %s\n", path);
+        return NS_INVALID_SOCKET;
+    }
 
     memset(&addr, 0, sizeof(addr));
     addr.sun_family = AF_UNIX;
-    strncpy(addr.sun_path, path, sizeof(addr.sun_path) - 1);
+    memcpy(addr.sun_path, path, pathLength + 1);
     unlink(path);
 
     sock = socket(AF_UNIX, socktype > 0 ? socktype : SOCK_STREAM, 0);
